Upgraded to v1.1
The library now does all the json encoding and decoding for you.

There is a new $token_file argument when you initiate the class. This
will be used as the full path to the file where the github oauth token
will be stored for future use.

Fixed the post_request method so that it now properly encodes and
passes the data through.